parse: refactor error handling and field types
Previously we we generating a `HashMap<String, String>` to make a map from
the group name to its URL. This is fine, but it results in the allocation
of a relatively larger data structure for little gain, because we can simply
make this cheap operation happen while looping through the groups. To that
end, we now only create a `Vec<String>` as the result of package parsing and
offload the URL generation to the actual downloading since that's the only
time it will be needed.

While at it, we also get rid of most `expect`/`unwrap` calls and pass errors
up the chain, attaching context through `anyhow::Context` where necessary.

Signed-off-by: Harsh Shandilya <me@msfjarvis.dev>